feat:  level 2,3
환전요청 CRUD
CustomError 예외처리

+++ src/main/java/com/sparta/currency_user/dto/exchange/ExchangeRequestDto.java
@@ -3,6 +3,8 @@ package com.sparta.currency_user.dto.exchange;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class ExchangeRequestDto {
 
@@ -10,9 +12,9 @@ public class ExchangeRequestDto {
 
     private final Long currencyId;
 
-    private final Long amountInKrw;
+    private final BigDecimal amountInKrw;
 
-    public ExchangeRequestDto(Long userId, Long currencyId, Long amountInKrw) {
+    public ExchangeRequestDto(Long userId, Long currencyId, BigDecimal amountInKrw) {
         this.userId = userId;
         this.currencyId = currencyId;
         this.amountInKrw = amountInKrw;
